// packages/tailwindcss-react-table/src/TailwindCSSReactTable.types.ts
type Prettify<T> = { [K in keyof T]: T[K] } & {};

type LiteralUnion<T extends U, U = string> = T | (U & Record<never, never>);

export type TRT_DensityState = 'comfortable' | 'compact' | 'spacious';

export interface TRT_Localization {
  actions: string;
  and: string;
  cancel: string;
  changeFilterMode: string;
  changeSearchMode: string;
  clearFilter: string;
  clearSearch: string;
  clearSort: string;
  clickToCopy: string;
  collapse: string;
  collapseAll: string;
  columnActions: string;
  copiedToClipboard: string;
  dropToGroupBy: string;
  edit: string;
  expand: string;
  expandAll: string;
  filterArrIncludes: string;
  filterArrIncludesAll: string;
  filterArrIncludesSome: string;
  filterBetween: string;
  filterBetweenInclusive: string;
  filterByColumn: string;
  filterContains: string;
  filterEmpty: string;
  filterEndsWith: string;
  filterEquals: string;
  filterEqualsString: string;
  filterFuzzy: string;
  filterGreaterThan: string;
  filterGreaterThanOrEqualTo: string;
  filterInNumberRange: string;
  filterIncludesString: string;
  filterIncludesStringSensitive: string;
  filterLessThan: string;
  filterLessThanOrEqualTo: string;
  filterMode: string;
  filterNotEmpty: string;
  filterNotEquals: string;
  filterStartsWith: string;
  filterWeakEquals: string;
  filteringByColumn: string;
  goToFirstPage: string;
  goToLastPage: string;
  goToNextPage: string;
  goToPreviousPage: string;
  grab: string;
  groupByColumn: string;
  groupedBy: string;
  hideAll: string;
  hideColumn: string;
  max: string;
  min: string;
  move: string;
  noRecordsToDisplay: string;
  noResultsFound: string;
  of: string;
  or: string;
  pinToLeft: string;
  pinToRight: string;
  resetColumnSize: string;
  resetOrder: string;
  rowActions: string;
  rowNumber: string;
  rowNumbers: string;
  rowsPerPage: string;
  save: string;
  search: string;
  select: string;
  selectedCountOfRowCountRowsSelected: string;
  showAll: string;
  showAllColumns: string;
  showHideColumns: string;
  showHideFilters: string;
  showHideSearch: string;
  sortByColumnAsc: string;
  sortByColumnDesc: string;
  sortedByColumnAsc: string;
  sortedByColumnDesc: string;
  thenBy: string;
  toggleDensity: string;
  toggleFullScreen: string;
  toggleSelectAll: string;
  toggleSelectRow: string;
  toggleVisibility: string;
  ungroupByColumn: string;
  unpin: string;
  unpinAll: string;
  unsorted: string;
}

export interface TRT_RowModel<TData extends Record<string, any> = {}> {
  flatRows: TRT_Row<TData>[];
  rows: TRT_Row<TData>[];
  rowsById: { [key: string]: TRT_Row<TData> };
}

export type TRT_TableInstance<TData extends Record<string, any> = {}> =
  Prettify<
    Omit<
      Table<TData>,
      | 'getAllColumns'
      | 'getAllFlatColumns'
      | 'getAllLeafColumns'
      | 'getCenterLeafColumns'
      | 'getColumn'
      | 'getExpandedRowModel'
      | 'getFlatHeaders'
      | 'getLeftLeafColumns'
      | 'getPaginationRowModel'
      | 'getPreFilteredRowModel'
      | 'getPrePaginationRowModel'
      | 'getRightLeafColumns'
      | 'getRowModel'
      | 'getSelectedRowModel'
      | 'getState'
      | 'options'
    > & {
      getAllColumns: () => TRT_Column<TData>[];
      getAllFlatColumns: () => TRT_Column<TData>[];
      getAllLeafColumns: () => TRT_Column<TData>[];
      getCenterLeafColumns: () => TRT_Column<TData>[];
      getColumn: (columnId: string) => TRT_Column<TData>;
      getExpandedRowModel: () => TRT_RowModel<TData>;
      getFlatHeaders: () => TRT_Header<TData>[];
      getLeftLeafColumns: () => TRT_Column<TData>[];
      getPaginationRowModel: () => TRT_RowModel<TData>;
      getPreFilteredRowModel: () => TRT_RowModel<TData>;
      getPrePaginationRowModel: () => TRT_RowModel<TData>;
      getRightLeafColumns: () => TRT_Column<TData>[];
      getRowModel: () => TRT_RowModel<TData>;
      getSelectedRowModel: () => TRT_RowModel<TData>;
      getState: () => TRT_TableState<TData>;
      options: TailwindCSSReactTableProps<TData> & {
        icons: TRT_Icons;
        localization: TRT_Localization;
      };
      refs: {
        bottomToolbarRef: MutableRefObject<HTMLDivElement>;
        editInputRefs: MutableRefObject<Record<string, HTMLInputElement>>;
        filterInputRefs: MutableRefObject<Record<string, HTMLInputElement>>;
        searchInputRef: MutableRefObject<HTMLInputElement>;
        tableContainerRef: MutableRefObject<HTMLDivElement>;
        tableHeadCellRefs: MutableRefObject<
          Record<string, HTMLTableCellElement>
        >;
        tablePaperRef: MutableRefObject<HTMLDivElement>;
        topToolbarRef: MutableRefObject<HTMLDivElement>;
      };
      setColumnFilterFns: Dispatch<
        SetStateAction<{ [key: string]: TRT_FilterOption }>
      >;
      setDensity: Dispatch<SetStateAction<TRT_DensityState>>;
      setDraggingColumn: Dispatch<SetStateAction<TRT_Column<TData> | null>>;
      setDraggingRow: Dispatch<SetStateAction<TRT_Row<TData> | null>>;
      setEditingCell: Dispatch<SetStateAction<TRT_Cell<TData> | null>>;
      setEditingRow: Dispatch<SetStateAction<TRT_Row<TData> | null>>;
      setGlobalFilterFn: Dispatch<SetStateAction<TRT_FilterOption>>;
      setHoveredColumn: Dispatch<
        SetStateAction<TRT_Column<TData> | { id: string } | null>
      >;
      setHoveredRow: Dispatch<
        SetStateAction<TRT_Row<TData> | { id: string } | null>
      >;
      setIsFullScreen: Dispatch<SetStateAction<boolean>>;
      setShowAlertBanner: Dispatch<SetStateAction<boolean>>;
      setShowColumnFilters: Dispatch<SetStateAction<boolean>>;
      setShowGlobalFilter: Dispatch<SetStateAction<boolean>>;
      setShowToolbarDropZone: Dispatch<SetStateAction<boolean>>;
    }
  >;

export type TRT_TableState<TData extends Record<string, any> = {}> = Prettify<
  TableState & {
    columnFilterFns: Record<string, TRT_FilterOption>;
    density: TRT_DensityState;
    draggingColumn: TRT_Column<TData> | null;
    draggingRow: TRT_Row<TData> | null;
    editingCell: TRT_Cell<TData> | null;
    editingRow: TRT_Row<TData> | null;
    globalFilterFn: TRT_FilterOption;
    hoveredColumn: TRT_Column<TData> | { id: string } | null;
    hoveredRow: TRT_Row<TData> | { id: string } | null;
    isFullScreen: boolean;
    isLoading: boolean;
    showAlertBanner: boolean;
    showColumnFilters: boolean;
    showGlobalFilter: boolean;
    showProgressBars: boolean;
    showSkeletons: boolean;
    showToolbarDropZone: boolean;
  }
>;
// ...
export type TRT_DefinedColumnDef<TData extends Record<string, any> = {}> =
  // Prettify<
  Omit<TRT_ColumnDef<TData>, 'id' | 'defaultDisplayColumn'> & {
    defaultDisplayColumn: Partial<TRT_ColumnDef<TData>>;
    id: string;
    _filterFn: TRT_FilterOption;
  };
// >

export type TRT_Column<TData extends Record<string, any> = {}> =
  // Prettify<
  Omit<
    Column<TData, unknown>,
    'header' | 'footer' | 'columns' | 'columnDef' | 'filterFn'
  > & {
    columnDef: TRT_DefinedColumnDef<TData>;
    columns?: TRT_Column<TData>[];
    filterFn?: TRT_FilterFn<TData>;
    footer: string;
    header: string;
  };
// >;

export type TRT_Header<TData extends Record<string, any> = {}> = Prettify<
  Omit<Header<TData, unknown>, 'column'> & {
    column: TRT_Column<TData>;
  }
>;

export type TRT_HeaderGroup<TData extends Record<string, any> = {}> = Prettify<
  Omit<HeaderGroup<TData>, 'headers'> & {
    headers: TRT_Header<TData>[];
  }
>;

export type TRT_Row<TData extends Record<string, any> = {}> = Prettify<
  Omit<
    Row<TData>,
    'getVisibleCells' | 'getAllCells' | 'subRows' | '_valuesCache'
  > & {
    getAllCells: () => TRT_Cell<TData>[];
    getVisibleCells: () => TRT_Cell<TData>[];
    subRows?: TRT_Row<TData>[];
    _valuesCache: Record<LiteralUnion<string & DeepKeys<TData>>, any>;
  }
>;

export type TRT_Cell<TData extends Record<string, any> = {}> = Prettify<
  Omit<Cell<TData, unknown>, 'column' | 'row'> & {
    column: TRT_Column<TData>;
    row: TRT_Row<TData>;
  }
>;
// ...
export type TRT_FilterOption = LiteralUnion<
  string & keyof typeof TRT_FilterFns
>;

export type TRT_FilterFn<TData extends Record<string, any> = {}> =
  | FilterFn<TData>
  | TRT_FilterOption;

export type TRT_InternalFilterOption = {
  option: string;
  symbol: string;
  label: string;
  divider: boolean;
};

export type TRT_DisplayColumnIds =
  | 'mrt-row-actions'
  | 'mrt-row-drag'
  | 'mrt-row-expand'
  | 'mrt-row-numbers'
  | 'mrt-row-select';